moved `open` to search module
